requirements.txt
numpy
picamera2
aiogram
opencv-python
python-dotenv